moved transformations to own python class
